Add files via upload
IDE: Google Colaboratory

Google Colab is a web IDE for python that allows to write and execute Python code without specific configuration.
It is a cloud-based Jupyter notebook environment. It runs in your web browser and lets anyone with internet access experiment with machine learning and coding for artificial intelligence.